Move Object\Contact to Model\Contact
