+++ boot.php
@@ -27,8 +27,7 @@ use Friendica\Core\Config;
 use Friendica\Core\PConfig;
 use Friendica\Core\Worker;
 use Friendica\Database\DBM;
-use Friendica\Object\Contact;
-use Friendica\Util\Lock;
+use Friendica\Model\Contact;
 
 require_once 'include/network.php';
 require_once 'include/plugin.php';
