+++ include/acl_selectors.php
@@ -6,8 +6,8 @@ use Friendica\App;
 use Friendica\Content\Feature;
 use Friendica\Core\Config;
 use Friendica\Database\DBM;
+use Friendica\Model\Contact;
 use Friendica\Model\GlobalContact;
-use Friendica\Object\Contact;
 
 require_once "include/contact_selectors.php";
 require_once "include/contact_widgets.php";
