+++ include/post_update.php
@@ -5,8 +5,8 @@
 
 use Friendica\Core\Config;
 use Friendica\Database\DBM;
+use Friendica\Model\Contact;
 use Friendica\Model\GlobalContact;
-use Friendica\Object\Contact;
 
 /**
  * @brief Calls the post update functions
