+++ mod/match.php
@@ -6,7 +6,7 @@ use Friendica\App;
 use Friendica\Core\Config;
 use Friendica\Core\System;
 use Friendica\Database\DBM;
-use Friendica\Object\Contact;
+use Friendica\Model\Contact;
 
 require_once 'include/text.php';
 require_once 'include/contact_widgets.php';
