+++ mod/photos.php
@@ -8,9 +8,9 @@ use Friendica\Core\System;
 use Friendica\Core\Config;
 use Friendica\Core\Worker;
 use Friendica\Database\DBM;
+use Friendica\Model\Contact;
 use Friendica\Model\Photo;
 use Friendica\Network\Probe;
-use Friendica\Object\Contact;
 use Friendica\Object\Image;
 
 require_once 'include/photos.php';
