+++ mod/unfollow.php
@@ -3,7 +3,7 @@
 use Friendica\App;
 use Friendica\Core\System;
 use Friendica\Database\DBM;
-use Friendica\Object\Contact;
+use Friendica\Model\Contact;
 
 require_once 'include/follow.php';
 require_once 'include/contact_selectors.php';
