+++ src/Protocol/OStatus.php
@@ -9,9 +9,9 @@ use Friendica\Core\Cache;
 use Friendica\Core\Config;
 use Friendica\Core\System;
 use Friendica\Database\DBM;
+use Friendica\Model\Contact;
 use Friendica\Model\GlobalContact;
 use Friendica\Network\Probe;
-use Friendica\Object\Contact;
 use Friendica\Object\Image;
 use Friendica\Util\Lock;
 use Friendica\Util\XML;
